feat: add GET user by id query refactor: move schemes to separate file
