+++ src/resources/users/users.schema.js
@@ -1,0 +1,60 @@
+// Options for GET user
+const getUserOpts = {
+  schema: {
+    response: {
+      200: {
+        type: 'array',
+        items: {
+          type: 'object',
+          properties: {
+            id: { type: 'string' },
+            name: { type: 'string' },
+            login: { type: 'string' },
+          },
+        },
+      },
+    },
+  },
+};
+
+// Options for POST user
+const postUserOpts = {
+  schema: {
+    body: {
+      type: 'object',
+      required: ['name', 'login', 'password'],
+      properties: {
+        name: { type: 'string' },
+        login: { type: 'string' },
+        password: { type: 'string' },
+      },
+    },
+    response: {
+      201: {
+        type: 'object',
+        properties: {
+          id: { type: 'string' },
+          name: { type: 'string' },
+          login: { type: 'string' },
+        },
+      },
+    },
+  },
+};
+
+// Options for GET users/:userId
+const getUserByIdOpts = {
+  schema: {
+    response: {
+      200: {
+        type: 'object',
+        properties: {
+          id: { type: 'string' },
+          name: { type: 'string' },
+          login: { type: 'string' },
+        },
+      },
+    },
+  },
+};
+module.exports = { getUserOpts, postUserOpts, getUserByIdOpts };
